# openapi.yaml
---
openapi: 3.0.2
info:
  title: FastAPI
  version: 0.1.0
paths:
  "/register":
    post:
      summary: Register
      operationId: register_register_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AuthSchema"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/login":
    post:
      summary: Login
      operationId: login_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              "$ref": "#/components/schemas/Body_login_login_post"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/users":
    get:
      summary: List Users
      operationId: list_users_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response List Users Users Get
                type: array
                items:
                  "$ref": "#/components/schemas/UserSchema"
  "/user/update":
    post:
      summary: Update User
      operationId: update_user_user_update_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserCanBeChangedSchema"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearer: []
  "/post":
    post:
      summary: Create Post
      operationId: create_post_post_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PostWithoutAuthor"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Post"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearer: []
  "/post/update":
    post:
      summary: Update Post
      operationId: update_post_post_update_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdatePost"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Post"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearer: []
  "/post/{post_id}":
    get:
      summary: Get Post
      operationId: get_post_post__post_id__get
      parameters:
      - required: true
        schema:
          title: Post Id
          type: string
        name: post_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Post
      operationId: delete_post_post__post_id__delete
      parameters:
      - required: true
        schema:
          title: Post Id
          type: string
        name: post_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearer: []
  "/posts":
    get:
      summary: Get All Posts
      operationId: get_all_posts_posts_get
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Cursor"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get All Posts Posts Get
                type: array
                items:
                  "$ref": "#/components/schemas/Post"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/like":
    get:
      summary: Like
      operationId: like_like_get
      parameters:
      - required: true
        schema:
          title: Post
          type: integer
        name: post
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearer: []
  "/view":
    get:
      summary: View
      operationId: view_view_get
      parameters:
      - required: true
        schema:
          title: Post
          type: integer
        name: post
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
      - OAuth2PasswordBearer: []
  "/post-stats":
    get:
      summary: Get Posts Stats
      operationId: get_posts_stats_post_stats_get
      parameters:
      - required: true
        schema:
          title: Post Id
          type: integer
        name: post_id
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PostStats"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/top-posts":
    get:
      summary: Get Top Posts
      operationId: get_top_posts_top_posts_get
      parameters:
      - required: true
        schema:
          title: Is Likes
          type: integer
        name: is_likes
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Top Posts Top Posts Get
                type: array
                items:
                  "$ref": "#/components/schemas/OnePostStat"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/top-users":
    get:
      summary: Get Top Posts
      operationId: get_top_posts_top_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Top Posts Top Users Get
                type: array
                items:
                  "$ref": "#/components/schemas/UserStats"
components:
  schemas:
    AuthSchema:
      title: AuthSchema
      required:
      - login
      - password
      type: object
      properties:
        login:
          title: Login
          type: string
        password:
          title: Password
          type: string
    Body_login_login_post:
      title: Body_login_login_post
      required:
      - username
      - password
      type: object
      properties:
        grant_type:
          title: Grant Type
          pattern: password
          type: string
        username:
          title: Username
          type: string
        password:
          title: Password
          type: string
        scope:
          title: Scope
          type: string
          default: ''
        client_id:
          title: Client Id
          type: string
        client_secret:
          title: Client Secret
          type: string
    Cursor:
      title: Cursor
      required:
      - login
      - cursor_start
      - cursor_end
      type: object
      properties:
        login:
          title: Login
          type: string
        cursor_start:
          title: Cursor Start
          type: integer
        cursor_end:
          title: Cursor End
          type: integer
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            "$ref": "#/components/schemas/ValidationError"
    OnePostStat:
      title: OnePostStat
      required:
      - post_id
      - author
      - stat
      type: object
      properties:
        post_id:
          title: Post Id
          type: integer
        author:
          title: Author
          type: string
        stat:
          title: Stat
          type: integer
    Post:
      title: Post
      required:
      - title
      - text
      - id
      - author
      - creationDate
      type: object
      properties:
        title:
          title: Title
          type: string
        text:
          title: Text
          type: string
        id:
          title: Id
          type: integer
        author:
          title: Author
          type: string
        creationDate:
          title: Creationdate
          type: integer
    PostStats:
      title: PostStats
      required:
      - post_id
      - likes
      - views
      type: object
      properties:
        post_id:
          title: Post Id
          type: integer
        likes:
          title: Likes
          type: integer
        views:
          title: Views
          type: integer
    PostWithoutAuthor:
      title: PostWithoutAuthor
      required:
      - title
      - text
      type: object
      properties:
        title:
          title: Title
          type: string
        text:
          title: Text
          type: string
    UpdatePost:
      title: UpdatePost
      required:
      - title
      - text
      - post_id
      type: object
      properties:
        title:
          title: Title
          type: string
        text:
          title: Text
          type: string
        post_id:
          title: Post Id
          type: integer
    UserCanBeChangedSchema:
      title: UserCanBeChangedSchema
      type: object
      properties:
        name:
          title: Name
          type: string
        surname:
          title: Surname
          type: string
        phone:
          title: Phone
          type: string
        email:
          title: Email
          type: string
        birthdate:
          title: Birthdate
          type: string
          format: date-time
    UserSchema:
      title: UserSchema
      required:
      - login
      - password
      type: object
      properties:
        login:
          title: Login
          type: string
        password:
          title: Password
          type: string
        name:
          title: Name
          type: string
        surname:
          title: Surname
          type: string
        phone:
          title: Phone
          type: string
        email:
          title: Email
          type: string
        birthdate:
          title: Birthdate
          type: string
          format: date-time
    UserStats:
      title: UserStats
      required:
      - login
      - likes
      type: object
      properties:
        login:
          title: Login
          type: string
        likes:
          title: Likes
          type: integer
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
            - type: string
            - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: login